refactor: centralize shared state and utilities
