+++ src/features/inventory/mutators.js
@@ -1,4 +1,4 @@
-import { S, save } from '../../game/state.js';
+import { S, save } from '../../shared/state.js';
 import { WEAPONS } from '../weaponGeneration/data/weapons.js';
 import { recomputePlayerTotals, canEquip } from './logic.js';
