+++ src/shared/mutators.js
@@ -1,0 +1,15 @@
+export * from '../features/ability/mutators.js';
+export * from '../features/adventure/mutators.js';
+export * from '../features/affixes/mutators.js';
+export * from '../features/alchemy/mutators.js';
+export * from '../features/combat/mutators.js';
+export * from '../features/cooking/mutators.js';
+export * from '../features/inventory/mutators.js';
+export * from '../features/karma/mutators.js';
+export * from '../features/loot/mutators.js';
+export * from '../features/mining/mutators.js';
+export * from '../features/physique/mutators.js';
+export * from '../features/proficiency/mutators.js';
+export * from '../features/progression/mutators.js';
+export * from '../features/sect/mutators.js';
+export * from '../features/weaponGeneration/mutators.js';
