+++ src/shared/selectors.js
@@ -1,0 +1,15 @@
+export * from '../features/ability/selectors.js';
+export * from '../features/adventure/selectors.js';
+export * from '../features/affixes/selectors.js';
+export * from '../features/alchemy/selectors.js';
+export * from '../features/combat/selectors.js';
+export * from '../features/cooking/selectors.js';
+export * from '../features/inventory/selectors.js';
+export * from '../features/karma/selectors.js';
+export * from '../features/loot/selectors.js';
+export * from '../features/mining/selectors.js';
+export * from '../features/physique/selectors.js';
+export * from '../features/proficiency/selectors.js';
+export * from '../features/progression/selectors.js';
+export * from '../features/sect/selectors.js';
+export * from '../features/weaponGeneration/selectors.js';
