+++ src/shared/utils/dom.js
@@ -1,4 +1,4 @@
-import { clamp } from '../features/progression/selectors.js';
+import { clamp } from '../../features/progression/selectors.js';
 
 export const qs = sel => document.querySelector(sel);
