+++ src/shared/utils/hp.js
