+++ ui/index.js
@@ -4,7 +4,7 @@
 // Way of Ascension — Modular JS
 
 import { LAWS } from '../src/features/progression/data/laws.js';
-import { S, defaultState, save, setState } from '../src/game/state.js';
+import { S, defaultState, save, setState } from '../src/shared/state.js';
 import {
   clamp,
   qCap,
@@ -28,7 +28,7 @@ import {
   getRealmName,
   checkLawUnlocks,
 } from '../src/features/progression/index.js';
-import { qs, setText, setFill, log } from '../src/game/utils.js';
+import { qs, setText, setFill, log } from '../src/shared/utils/dom.js';
 import { createProgressBar, updateProgressBar } from './components/progressBar.js';
 import { renderSidebarActivities } from '../src/ui/sidebar.js';
 import { initializeWeaponChip, updateWeaponChip } from '../src/features/inventory/ui/weaponChip.js';
